feat: restructure settings components and add PriceCalc and AddCities for improved user experience
