+++ components/order/orderTableView.tsx
@@ -49,10 +49,6 @@ export default function OrderTabelView({ redirectLink }: props) {
   const [filterValue, setFilterValue] = useState<string>('All Status');
   const [orderAmount, setOrderAmount] = useState<number>(0);
 
-  const role = user?.role;
-  const name = user?.name;
-  const userId = user?.id;
-
   useEffect(() => {
     const fetchOrders = async () => {
       const decoded = await userProfile();
@@ -65,7 +61,7 @@ export default function OrderTabelView({ redirectLink }: props) {
       }
       try {
         const response = await FetchOrders({
-          userid: user.id!,
+          userid: user!.id,
           pagenumber: pagenumber,
         });
 
@@ -322,7 +318,7 @@ export default function OrderTabelView({ redirectLink }: props) {
     }
   };
   const handleDelete = async (id: string) => {
-    const response = await DeleteOrder({ userid: userId!, trxCode: id });
+    const response = await DeleteOrder({ userid: user!.id, trxCode: id });
     if (response.message === 'Order deleted successfully') {
       toast({
         title: 'Deleted Successfully',
@@ -344,7 +340,7 @@ export default function OrderTabelView({ redirectLink }: props) {
       <div className="h-fit justify-start items-center gap-9 inline-flex">
         <div className="flex-col justify-start items-start gap-2 inline-flex">
           <div className="self-stretch text-[#060A87] text-2xl md:text-3xl font-extrabold font-['Manrope'] leading-[36px]">
-            Welcome Back, {name}!
+            Welcome Back, {user!.name}!
           </div>
           <div className="self-stretch text-[#495d85] text-sm md:text-base font-extrabold font-['Manrope'] leading-tight">
             Here’s your Orders Report
@@ -364,7 +360,7 @@ export default function OrderTabelView({ redirectLink }: props) {
           </button>
         </div>
         <AddOrderDialogue
-          userId={userId!}
+          userId={user!.id}
           showNewOrderModal={showNewOrderModal}
           setShowNewOrderModal={setShowNewOrderModal}
           showConfirmationModal={showConfirmationModal}
@@ -388,7 +384,7 @@ export default function OrderTabelView({ redirectLink }: props) {
             setCurrentPage={setCurrentPage}
             pagenumber={pagenumber}
             setPagenumber={setPagenumber}
-            role={role!}
+            role={user!.role}
             name={name!}
             columns={
               columns as ColumnDef<
@@ -414,7 +410,7 @@ export default function OrderTabelView({ redirectLink }: props) {
             setCurrentPage={setCurrentPage}
             pagenumber={pagenumber}
             setPagenumber={setPagenumber}
-            role={role!}
+            role={user!.role}
             name={name!}
             columns={
               columns as ColumnDef<
